sim proj

+++ modules/module1/include/framepoint.h
@@ -8,14 +8,14 @@ namespace Data
     class FramePoint
     {
         protected:
-            std::vector<cv::Point3f> mvFramePoints3d;
+            std::vector<cv::Point3d> mvFramePoints3d;
             std::vector<cv::KeyPoint> mvFramePoints2d;
 
         public:
             FramePoint() = default;
 
-            void setFramePoint3d(std::vector<cv::Point3f> framePoints);
-            std::vector<cv::Point3f> getFramePoint3d();
+            void setFramePoint3d(std::vector<cv::Point3d> framePoints);
+            std::vector<cv::Point3d> getFramePoint3d();
 
 
             void setFramePoint2d(std::vector<cv::KeyPoint> framePoints);
